--- src/constants/teams.ts
@@ -1,27 +1,0 @@
-type TeamCodeType = {
-  [key: string]: string;
-};
-
-export const TeamCode: TeamCodeType = {
-  ars: "Arsenal",
-  avl: "Aston Villa",
-  bou: "Bournemouth",
-  bre: "Brentford",
-  bha: "Brighton & Hove Albion",
-  bur: "Burnley",
-  che: "Chelsea",
-  cry: "Crystal Palace",
-  eve: "Everton",
-  ful: "Fulham",
-  liv: "Liverpool",
-  lut: "Luton Town",
-  mci: "Manchester City",
-  mun: "Manchester United",
-  new: "Newcastle United",
-  nfo: "Nottingham Forest",
-  shu: "Sheffield United",
-  tot: "Tottenham Hotspur",
-  whu: "West Ham United",
-  wol: "Wolverhampton Wanderers",
-};
-export type TeamCodeKey = keyof typeof TeamCode;
