+++ src/types/soccer/team.ts
